feat: add Silkscreen-Regular.woff font file and update font-face declaration in Layout.astro
The Silkscreen-Regular.woff font file was added to the public/fonts directory. The font-face declaration in the Layout.astro file was updated to include the Silkscreen-Regular.woff font file in addition to the existing Silkscreen-Regular.ttf font file. This allows the browser to use the WOFF font format if available, providing better font support.